subcase_1b/training_platform/requirements.txt
Flask
requests
PyJWT
Jinja2
PyYAML
